pull track into separate file
