+++ src/lib/constants.js
@@ -1,4 +1,4 @@
-{
+module.exports = {
   "sevenDays": 6.048e+8,
   "accessTokenUrl": "https://www.paypal.com/muse/api/partner-token",
   "storage": {
