+++ src/lib/cookie-utils.js
@@ -1,4 +1,5 @@
 /* @flow */
+import generate from './generate-id';
 
 export const getCookie = (cookieName : string) : string => {
     const name = `${ cookieName }=`;
@@ -20,4 +21,13 @@ export const setCookie = (cookieName : string, cookieValue : string, expirationM
     d.setTime(d.getTime() + expirationMilliseconds);
     const expires = `expires=${ d.toUTCString() }`;
     document.cookie = `${ cookieName }=${ cookieValue }; Path=/; ${ expires }`;
+};
+
+export const getUserIdCookie = () : ?string => {
+    return getCookie('paypal-user-id') || null;
+};
+
+export const setRandomUserIdCookie = () : void => {
+    const ONE_MONTH_IN_MILLISECONDS = 30 * 24 * 60 * 60 * 1000;
+    setCookie('paypal-user-id', generate.generateId(), ONE_MONTH_IN_MILLISECONDS);
 };
